+++ MyApplication/app/src/main/java/com/example/overmind/SimulationParameters.java
@@ -2,13 +2,13 @@ package com.example.overmind;
 
 class SimulationParameters {
 
-    static private double recoveryScale;
-    static private double recoverySensitivity;
-    static private double resetPotential;
-    static private double resetRecovery;
-    static private double current;
+    static private float recoveryScale;
+    static private float recoverySensitivity;
+    static private float resetPotential;
+    static private float resetRecovery;
+    static private float current;
 
-    synchronized  static void setParameters(double a, double b, double c, double d, double i) {
+    synchronized  static void setParameters(float a, float b, float c, float d, float i) {
         recoveryScale = a;
         recoverySensitivity = b;
         resetPotential = c;
@@ -16,9 +16,9 @@ class SimulationParameters {
         current = i;
     }
 
-    synchronized static double[] getParameters() {
+    synchronized static float[] getParameters() {
 
-        double[] paramArray = new double[5];
+        float[] paramArray = new float[5];
 
         paramArray[0] = recoveryScale;
         paramArray[1] = recoverySensitivity;
